+++ elements/src/components/pos-predicate/pos-predicate.tsx
@@ -1,0 +1,16 @@
+import { Component, h, Prop } from '@stencil/core';
+
+@Component({
+  tag: 'pos-predicate',
+})
+export class PosPredicate {
+  @Prop()
+  uri: string;
+
+  @Prop()
+  label: string;
+
+  render() {
+    return <ion-label title={this.uri}>{this.label}</ion-label>;
+  }
+}